move version configuration to client implementation
