+++ reactor-netty-http/src/main/java/reactor/netty/http/client/WebsocketClientSpec.java
@@ -15,7 +15,10 @@
  */
 package reactor.netty.http.client;
 
+import io.netty.handler.codec.http.websocketx.WebSocketVersion;
 import reactor.netty.http.websocket.WebsocketSpec;
+
+import java.util.Objects;
 
 /**
  * Websocket client configuration
@@ -26,8 +29,16 @@ import reactor.netty.http.websocket.WebsocketSpec;
 public interface WebsocketClientSpec extends WebsocketSpec {
 
 	/**
+	 * Returns the configured WebSocket version
+	 *
+	 * @return returns the configured WebSocket version
+	 * @since 1.0.3
+	 */
+	WebSocketVersion version();
+
+	/**
 	 * Create builder with default properties:<br>
-	 * version = 13
+	 * version = {@link io.netty.handler.codec.http.websocketx.WebSocketVersion#V13}
 	 * <br>
 	 * protocols = null
 	 * <br>
@@ -44,8 +55,28 @@ public interface WebsocketClientSpec extends WebsocketSpec {
 	}
 
 	final class Builder extends WebsocketSpec.Builder<Builder> {
+		WebSocketVersion version = WebSocketVersion.V13;
 
 		private Builder() {
+		}
+
+		/**
+		 * Sets websocket version to use.
+		 * Set to {@link io.netty.handler.codec.http.websocketx.WebSocketVersion#V13} by default
+		 *
+		 * @param version WebSocket version to be used
+		 * @return {@literal this}
+		 * @throws NullPointerException if version is null
+		 * @throws IllegalArgumentException if the version is unknown
+		 * @since 1.0.3
+		 */
+		public final Builder version(WebSocketVersion version) {
+			Objects.requireNonNull(version, "version");
+			if (WebSocketVersion.UNKNOWN.equals(version)) {
+				throw new IllegalArgumentException("Unknown web socket version:" + version);
+			}
+			this.version = version;
+			return this;
 		}
 
 		/**
