+++ reactor-netty-http/src/main/java/reactor/netty/http/client/WebsocketClientSpecImpl.java
@@ -15,6 +15,7 @@
  */
 package reactor.netty.http.client;
 
+import io.netty.handler.codec.http.websocketx.WebSocketVersion;
 import reactor.netty.http.websocket.WebsocketSpecImpl;
 
 /**
@@ -22,8 +23,15 @@ import reactor.netty.http.websocket.WebsocketSpecImpl;
  * @since 0.9.7
  */
 final class WebsocketClientSpecImpl extends WebsocketSpecImpl implements WebsocketClientSpec {
+	private final WebSocketVersion version;
+
+	@Override
+	public WebSocketVersion version() {
+		return version;
+	}
 
 	WebsocketClientSpecImpl(WebsocketClientSpec.Builder builder) {
 		super(builder);
+		this.version = builder.version;
 	}
 }
